extract common logic out of CreateModule.kt into ReplaceInFiles.kt
