+++ buildSrc/src/main/kotlin/CreateModule.kt
@@ -1,16 +1,11 @@
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.TaskAction
 import java.io.File
-import java.nio.file.FileVisitResult
-import java.nio.file.Files
-import java.nio.file.Path
-import java.nio.file.SimpleFileVisitor
-import java.nio.file.attribute.BasicFileAttributes
 
-val skeletonModuleName = "_skeleton"
-val skeletonJavaModuleName = "_skeleton-java"
-val languageInputPrompt = "Which language do you want to use? [kotlin, java] : "
-val moduleInputPrompt = "Enter module name ( ${moduleNameRegex.pattern} ) : "
+private const val skeletonModuleName = "_skeleton"
+private const val skeletonJavaModuleName = "_skeleton-java"
+private const val languageInputPrompt = "Which language do you want to use? [kotlin, java] : "
+private val moduleInputPrompt = "Enter module name ( ${moduleNameRegex.pattern} ) : "
 
 open class CreateModule : DefaultTask() {
 
@@ -61,77 +56,8 @@ open class CreateModule : DefaultTask() {
 
     }
 
-    fun copySkeletonTo(skeleton: String, target: File) {
+    private fun copySkeletonTo(skeleton: String, target: File) {
         val skeletonDir = File(project.projectDir, skeleton)
         skeletonDir.copyRecursively(target)
     }
-
-    fun replaceInFilesAndFilenames(rootDirectory: File, replacements: Map<String, String>) {
-        if (!rootDirectory.exists() || !rootDirectory.isDirectory) {
-            println("The provided path does not exist or is not a directory.")
-            return
-        }
-
-        // Separate the processing into two phases: content replacement and renaming
-        try {
-            // Phase 1: Replace contents without renaming files/directories yet
-            Files.walkFileTree(rootDirectory.toPath(), object : SimpleFileVisitor<Path>() {
-                override fun visitFile(file: Path, attrs: BasicFileAttributes): FileVisitResult {
-                    replaceInFileContent(file.toFile(), replacements)
-                    return FileVisitResult.CONTINUE
-                }
-            })
-
-            // Phase 2: Collect and rename files/directories, starting from the deepest part
-            val pathsToRename = mutableListOf<Path>()
-            Files.walkFileTree(rootDirectory.toPath(), object : SimpleFileVisitor<Path>() {
-                override fun preVisitDirectory(dir: Path, attrs: BasicFileAttributes): FileVisitResult {
-                    pathsToRename.add(dir)
-                    return FileVisitResult.CONTINUE
-                }
-
-                override fun visitFile(file: Path, attrs: BasicFileAttributes): FileVisitResult {
-                    pathsToRename.add(file)
-                    return FileVisitResult.CONTINUE
-                }
-            })
-
-            // Reverse the list to rename deepest files/directories first
-            pathsToRename.reverse()
-            for (path in pathsToRename) {
-                replaceInFilename(path.toFile(), replacements)
-            }
-        } catch (e: Exception) {
-            println("An error occurred while processing the directory: ${e.message}")
-        }
-    }
-
-    fun replaceInFileContent(file: File, replacements: Map<String, String>) {
-        try {
-            var content = file.readText()
-            replacements.forEach { (key, value) ->
-                content = content.replace("%{$key}%", value)
-            }
-            file.writeText(content)
-        } catch (e: Exception) {
-            println("Failed to replace content in file: ${file.path}. Error: ${e.message}")
-        }
-    }
-
-    fun replaceInFilename(file: File, replacements: Map<String, String>): String {
-        var newName = file.name
-        replacements.forEach { (key, value) ->
-            newName = newName.replace("%{$key}%", value)
-        }
-        if (newName != file.name) {
-            val newFile = File(file.parent, newName)
-            if (file.renameTo(newFile)) {
-                return newFile.path
-            } else {
-                println("Failed to rename file: ${file.path} to $newName")
-            }
-        }
-        return file.path
-    }
-
 }
